Sets argument checking on/off "semi-globally"

You can allow your users to turn off argument checking within your function by using set_check. Only the functions check_arg nd check_value can be turned off that way.

set_check(x)

Arguments

x

A logical scalar, no default.

Details

This function can be useful if you develop a function that may be used in large range loops (>100K). In such situations, it may be good to still check all arguments, but to offer the user to turn this checking off with an extra argument (named arg.check for instance). Doing so you would achieve the feat of i) having a user-friendly function thanks to argument checking and, ii) still achieve high performance in large loops (although the computational footprint of argument checking is quite low around 30 micro seconds for missing arguments to 80 micro seconds for non-missing arguments of simple type).

Examples

# Let's give an example
test_check = function(x, y, arg.check = TRUE){
  set_check(arg.check)
  check_arg(x, y, "numeric scalar")
  x + y
}

# Works: argument checking on
test_check(1, 2)
#> [1] 3

# If mistake, nice error msg
try(test_check(1, "a"))
#> Error : in test_check(1, "a"):
#>  Argument `y` must be a numeric scalar. Problem: it is not numeric (instead
#>  it is of type 'character').

# Now argument checking turned off
test_check(1, 2, FALSE)
#> [1] 3
# But if mistake: "not nice" error message
try(test_check(1, "a", FALSE))
#> Error in x + y : non-numeric argument to binary operator